+++ event-sourcing/event-repository/event-repository-jdbc/src/main/java/uk/gov/justice/services/eventsourcing/repository/jdbc/discovery/StreamPosition.java
@@ -1,0 +1,6 @@
+package uk.gov.justice.services.eventsourcing.repository.jdbc.discovery;
+
+import java.util.UUID;
+
+public record StreamPosition(UUID streamId, Long positionInStream) {
+}